requirements.txt
commonroad-io==2020.3
commonroad-drivability-checker
commonroad-helper-functions==1.0.0
commonroad-agent==1.1.0
wale-net>=2.0.1
numpy~=1.19.5
matplotlib~=3.2.2
progressbar2
argparse~=1.4.0
PyPDF2~=1.26.0
joblib~=0.16.0
scipy~=1.5.0

shapely~=1.7.1
numba~=0.50.1
tqdm~=4.47.0
setuptools~=49.2.0
networkx==2.5